add id code to same file, refactor writer
