+++ pkg/resources/task.go
@@ -1,21 +1,14 @@
 package resources
 
 import (
-	"bytes"
 	"database/sql"
-	"encoding/csv"
 	"fmt"
 	"log"
-	"strings"
 
 	"github.com/chanzuckerberg/terraform-provider-snowflake/pkg/snowflake"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/pkg/errors"
-)
-
-const (
-	taskIDDelimiter = '|'
 )
 
 var taskSchema = map[string]*schema.Schema{
@@ -87,27 +80,6 @@ var taskSchema = map[string]*schema.Schema{
 		Description: "Any single SQL statement, or a call to a stored procedure, executed when the task runs.",
 		ForceNew:    false,
 	},
-}
-
-type taskID struct {
-	DatabaseName string
-	SchemaName   string
-	TaskName     string
-}
-
-//String() takes in a taskID object and returns a pipe-delimited string:
-//DatabaseName|SchemaName|TaskName
-func (t *taskID) String() (string, error) {
-	var buf bytes.Buffer
-	csvWriter := csv.NewWriter(&buf)
-	csvWriter.Comma = taskIDDelimiter
-	dataIdentifiers := [][]string{{t.DatabaseName, t.SchemaName, t.TaskName}}
-	err := csvWriter.WriteAll(dataIdentifiers)
-	if err != nil {
-		return "", err
-	}
-	strTaskID := strings.TrimSpace(buf.String())
-	return strTaskID, nil
 }
 
 // difference find keys in a but not in b
@@ -200,31 +172,6 @@ func resumeTask(root *snowflake.TaskBuilder, meta interface{}) {
 	if err != nil {
 		log.Fatal(errors.Wrapf(err, "error resuming root task %v", root.QualifiedName()))
 	}
-}
-
-// taskIDFromString() takes in a pipe-delimited string: DatabaseName|SchemaName|TaskName
-// and returns a taskID object
-func taskIDFromString(stringID string) (*taskID, error) {
-	reader := csv.NewReader(strings.NewReader(stringID))
-	reader.Comma = pipeIDDelimiter
-	lines, err := reader.ReadAll()
-	if err != nil {
-		return nil, fmt.Errorf("Not CSV compatible")
-	}
-
-	if len(lines) != 1 {
-		return nil, fmt.Errorf("1 line per task")
-	}
-	if len(lines[0]) != 3 {
-		return nil, fmt.Errorf("3 fields allowed")
-	}
-
-	taskResult := &taskID{
-		DatabaseName: lines[0][0],
-		SchemaName:   lines[0][1],
-		TaskName:     lines[0][2],
-	}
-	return taskResult, nil
 }
 
 // Task returns a pointer to the resource representing a task
